+++ SME.CDEP.Aplicacao/Servicos/ServicoImportacaoArquivoAcervoAudiovisual.cs
@@ -8,6 +8,7 @@ using SME.CDEP.Dominio.Constantes;
 using SME.CDEP.Dominio.Entidades;
 using SME.CDEP.Dominio.Excecoes;
 using SME.CDEP.Dominio.Extensions;
+using SME.CDEP.Infra;
 using SME.CDEP.Infra.Dados.Repositorios.Interfaces;
 using SME.CDEP.Infra.Dominio.Enumerados;
 
@@ -15,24 +16,17 @@ namespace SME.CDEP.Aplicacao.Servicos
 {
     public class ServicoImportacaoArquivoAcervoAudiovisual : ServicoImportacaoArquivoBase, IServicoImportacaoArquivoAcervoAudiovisual 
     {
-        private readonly IServicoAcervoAudiovisual servicoAcervoAudiovisual;
-        private readonly IMapper mapper;
+        private readonly IServicoMensageria servicoMensageria;
         
         public ServicoImportacaoArquivoAcervoAudiovisual(IRepositorioImportacaoArquivo repositorioImportacaoArquivo, IServicoMaterial servicoMaterial,
             IServicoEditora servicoEditora,IServicoSerieColecao servicoSerieColecao,IServicoIdioma servicoIdioma, IServicoAssunto servicoAssunto,
             IServicoCreditoAutor servicoCreditoAutor,IServicoConservacao servicoConservacao, IServicoAcessoDocumento servicoAcessoDocumento,
-            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte,IServicoFormato servicoFormato,IServicoAcervoAudiovisual servicoAcervoAudiovisual, IMapper mapper,
+            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte,IServicoFormato servicoFormato,IServicoMensageria servicoMensageria, IMapper mapper,
             IRepositorioParametroSistema repositorioParametroSistema)
             : base(repositorioImportacaoArquivo, servicoMaterial, servicoEditora,servicoSerieColecao, servicoIdioma, servicoAssunto, servicoCreditoAutor,
                 servicoConservacao,servicoAcessoDocumento,servicoCromia,servicoSuporte, servicoFormato, mapper, repositorioParametroSistema)
         {
-            this.servicoAcervoAudiovisual = servicoAcervoAudiovisual ?? throw new ArgumentNullException(nameof(servicoAcervoAudiovisual));
-            this.mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
-        }
-
-        public void DefinirCreditosAutores(List<IdNomeTipoDTO> creditosAutores)
-        {
-            CreditosAutores = creditosAutores;
+            this.servicoMensageria = servicoMensageria ?? throw new ArgumentNullException(nameof(servicoMensageria));
         }
         
         public async Task<bool> RemoverLinhaDoArquivo(long id, LinhaDTO linhaDoArquivo)
@@ -74,15 +68,6 @@ namespace SME.CDEP.Aplicacao.Servicos
             return await ObterRetornoImportacaoAcervo(arquivoImportado, JsonConvert.DeserializeObject<IEnumerable<AcervoAudiovisualLinhaDTO>>(arquivoImportado.Conteudo), false);
         }
 
-        public async Task CarregarDominiosAudiovisuais()
-        {
-            await CarregarTodosOsDominios();
-            
-            await ObterSuportesPorTipo(TipoSuporte.VIDEO);
-            
-            await ObterCreditosAutoresPorTipo(TipoCreditoAutoria.Credito);
-        }
-
         public async Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoAudiovisualDTO,AcervoAudiovisualLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ImportarArquivo(IFormFile file)
         {
             ValidarArquivo(file);
@@ -91,18 +76,18 @@ namespace SME.CDEP.Aplicacao.Servicos
         
             var importacaoArquivo = ObterImportacaoArquivoParaSalvar(file.FileName, TipoAcervo.Audiovisual, JsonConvert.SerializeObject(acervosAudiovisualLinhas));
 
-            await CarregarDominiosAudiovisuais();
-            
             var importacaoArquivoId = await PersistirImportacao(importacaoArquivo);
            
-            ValidarPreenchimentoValorFormatoQtdeCaracteres(acervosAudiovisualLinhas);
-            
-            await PersistenciaAcervo(acervosAudiovisualLinhas);
-            await AtualizarImportacao(importacaoArquivoId, JsonConvert.SerializeObject(acervosAudiovisualLinhas), acervosAudiovisualLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
-        
-            var arquivoImportado = await repositorioImportacaoArquivo.ObterPorId(importacaoArquivoId);
-        
-            return await ObterRetornoImportacaoAcervo(arquivoImportado, acervosAudiovisualLinhas);
+            await servicoMensageria.Publicar(RotasRabbit.ExecutarImportacaoArquivoAcervoAudiovisual, importacaoArquivoId, Guid.NewGuid());
+
+            return new ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoAudiovisualDTO,AcervoAudiovisualLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>()
+            {
+                Id = importacaoArquivoId,
+                Nome = importacaoArquivo.Nome,
+                TipoAcervo = importacaoArquivo.TipoAcervo,
+                DataImportacao = DateTimeExtension.HorarioBrasilia(),
+                Status = ImportacaoStatus.Pendente
+            };
         }
         
         private async Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoAudiovisualDTO,AcervoAudiovisualLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ObterRetornoImportacaoAcervo(ImportacaoArquivo arquivoImportado, IEnumerable<AcervoAudiovisualLinhaDTO> acervosAudiovisualLinhas, bool estaImportandoArquivo = true)
@@ -247,113 +232,6 @@ namespace SME.CDEP.Aplicacao.Servicos
                 mensagemErro.Add(acervoAudiovisualLinhaDTO.Disponibilizacao.Mensagem);
 
             return mensagemErro.ToArray();
-        }
-        
-        public async Task PersistenciaAcervo(IEnumerable<AcervoAudiovisualLinhaDTO> acervosAudiovisualLinhas)
-        {
-            foreach (var acervoAudiovisualLinha in acervosAudiovisualLinhas.Where(w=> !w.PossuiErros))
-            {
-                try
-                {
-                    var acervoAudiovisual = new AcervoAudiovisualCadastroDTO()
-                    {
-                        Titulo = acervoAudiovisualLinha.Titulo.Conteudo,
-                        Codigo = acervoAudiovisualLinha.Codigo.Conteudo,
-                        CreditosAutoresIds = CreditosAutores
-                            .Where(f => acervoAudiovisualLinha.Credito.Conteudo.RemoverAcentuacaoFormatarMinusculo().FormatarTextoEmArray().Contains(f.Nome.RemoverAcentuacaoFormatarMinusculo()))
-                            .Select(s => s.Id).ToArray(),
-                        Localizacao = acervoAudiovisualLinha.Localizacao.Conteudo,
-                        Procedencia = acervoAudiovisualLinha.Procedencia.Conteudo,
-                        Ano = acervoAudiovisualLinha.Ano.Conteudo,
-                        Copia = acervoAudiovisualLinha.Copia.Conteudo,
-                        PermiteUsoImagem = ObterAutorizaUsoDeImagemPorValorDoCampo(acervoAudiovisualLinha.PermiteUsoImagem.Conteudo),
-                        ConservacaoId = acervoAudiovisualLinha.EstadoConservacao.Conteudo.EstaPreenchido() ? ObterConservacaoIdPorValorDoCampo(acervoAudiovisualLinha.EstadoConservacao.Conteudo) : null,
-                        Descricao = acervoAudiovisualLinha.Descricao.Conteudo,
-                        SuporteId = ObterSuporteVideoIdPorValorDoCampo(acervoAudiovisualLinha.Suporte.Conteudo),
-                        Duracao = acervoAudiovisualLinha.Duracao.Conteudo,
-                        CromiaId = acervoAudiovisualLinha.Cromia.Conteudo.EstaPreenchido() ? ObterCromiaIdPorValorDoCampo(acervoAudiovisualLinha.Cromia.Conteudo) : null,
-                        TamanhoArquivo = acervoAudiovisualLinha.TamanhoArquivo.Conteudo,
-                        Acessibilidade = acervoAudiovisualLinha.Acessibilidade.Conteudo,
-                        Disponibilizacao = acervoAudiovisualLinha.Disponibilizacao.Conteudo,
-                    };
-                    await servicoAcervoAudiovisual.Inserir(acervoAudiovisual);
-        
-                    acervoAudiovisualLinha.DefinirLinhaComoSucesso();
-                }
-                catch (Exception ex)
-                {
-                    acervoAudiovisualLinha.DefinirLinhaComoErro(ex.Message);
-                }
-            }
-        } 
-        
-        public void ValidarPreenchimentoValorFormatoQtdeCaracteres(IEnumerable<AcervoAudiovisualLinhaDTO> linhas)
-        {
-            var creditos = CreditosAutores.Where(w => w.Tipo == (int)TipoCreditoAutoria.Credito).Select(s=> mapper.Map<IdNomeDTO>(s));
-            var suportesDeVideo = Suportes.Where(w => w.Tipo == (int)TipoSuporte.VIDEO).Select(s=> mapper.Map<IdNomeDTO>(s));
-            
-            foreach (var linha in linhas)
-            {
-                try
-                {
-                    ValidarPreenchimentoLimiteCaracteres(linha.Titulo, Constantes.TITULO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Codigo, Constantes.TOMBO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Credito, Constantes.CREDITO);
-                    ValidarConteudoCampoListaComDominio(linha.Credito, creditos, Constantes.CREDITO);	
-
-                    ValidarPreenchimentoLimiteCaracteres(linha.Localizacao,Constantes.LOCALIZACAO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Procedencia,Constantes.PROCEDENCIA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Ano,Constantes.ANO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Copia,Constantes.COPIA);
-                    ValidarPreenchimentoLimiteCaracteres(linha.PermiteUsoImagem,Constantes.AUTORIZACAO_USO_DE_IMAGEM);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.EstadoConservacao,Constantes.ESTADO_CONSERVACAO);
-                    ValidarConteudoCampoComDominio(linha.EstadoConservacao, Conservacoes, Constantes.ESTADO_CONSERVACAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Descricao,Constantes.DESCRICAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Suporte,Constantes.SUPORTE);
-                    ValidarConteudoCampoComDominio(linha.Suporte, suportesDeVideo, Constantes.SUPORTE);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Duracao,Constantes.DURACAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Cromia,Constantes.CROMIA);
-                    ValidarConteudoCampoComDominio(linha.Cromia, Cromias, Constantes.CROMIA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.TamanhoArquivo,Constantes.TAMANHO_ARQUIVO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Acessibilidade,Constantes.ACESSIBILIDADE);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Disponibilizacao,Constantes.DISPONIBILIDADE);
-                    linha.PossuiErros = PossuiErro(linha);
-                }
-                catch (Exception e)
-                {
-                    linha.DefinirLinhaComoErro(string.Format(Constantes.OCORREU_UMA_FALHA_INESPERADA_NA_LINHA_X_MOTIVO_Y, linha.NumeroLinha, e.Message));
-                }
-            }
-        }
-
-        private bool PossuiErro(AcervoAudiovisualLinhaDTO linha)
-        {
-            return linha.Titulo.PossuiErro 
-                   || linha.Codigo.PossuiErro 
-                   || linha.Credito.PossuiErro
-                   || linha.Localizacao.PossuiErro 
-                   || linha.Procedencia.PossuiErro
-                   || linha.Ano.PossuiErro
-                   || linha.Copia.PossuiErro 
-                   || linha.PermiteUsoImagem.PossuiErro 
-                   || linha.EstadoConservacao.PossuiErro
-                   || linha.Descricao.PossuiErro
-                   || linha.Suporte.PossuiErro
-                   || linha.Duracao.PossuiErro
-                   || linha.Cromia.PossuiErro 
-                   || linha.TamanhoArquivo.PossuiErro 
-                   || linha.Acessibilidade.PossuiErro
-                   || linha.Disponibilizacao.PossuiErro;
         }
         
         private async Task<IEnumerable<AcervoAudiovisualLinhaDTO>> LerPlanilha(IFormFile file)
